requirements.txt
Django==2.2.9
djangorestframework==3.11.0
mysqlclient==1.4.6
requests==2.22.0
django-redis==4.11.0
django-filter==2.2.0
ipython==7.11.1
django-cas-ng==4.0.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
environs==7.3.0
celery==4.4.0
xlrd==1.2.0
XlsxWriter==1.2.8
django-mptt==0.11.0
drf-extensions==0.6.0
names==0.3.0
aiokafka==0.6.0
retrying==1.3.3
pykafka==2.8.0
pika==1.1.0
gevent==1.4.0
itsdangerous==1.1.0
suds-py3==1.4.1.0
reportlab==3.5.42
django-cluster-redis==1.0.5
xmltodict==0.12.0
PyYAML==5.3.1
pytest-django==3.9.0
pytest-rerunfailures==9.1
allure-pytest==2.8.18
allure-python-commons==2.8.18
redis-py-cluster==2.1.0